JS - Yelp Camp : Authentication - Campground Permissions

+++ 20.YelpCamp/routes/campground.js
@@ -67,6 +67,10 @@ router.get(
       req.flash('error', 'Cannot find that campground');
       return res.redirect('/campgrounds');
     }
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to do that!');
+      return res.redirect(`/campgrounds/${req.params.id}`);
+    }
     res.render('campgrounds/edit', { campground });
   })
 );
@@ -77,7 +81,13 @@ router.put(
   validateCampground,
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
+    // check to see does this campground have the same author ID as the currently logged in user?
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to do that!');
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
     req.flash('success', 'Successfully updated campground!');
     res.redirect(`/campgrounds/${campground._id}`);
   })
@@ -88,6 +98,11 @@ router.delete(
   isLoggedIn,
   catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to do that!');
+      return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted campground');
     res.redirect('/campgrounds');
